Beautify code using clang-code.

+++ pybindcpp/ext/simple.cpp
@@ -3,27 +3,13 @@
 
 using namespace pybindcpp;
 
-int
-g(int x, int y)
-{
-  return x + y;
-}
+int g(int x, int y) { return x + y; }
 
-int
-f(double x, double y)
-{
-  return x * y;
-}
+double f(double x, double y) { return x * y; }
 
-void
-simple(ExtModule& m)
-{
+void simple(ExtModule &m) {
   m.fun("g", g);
   m.fun("f", f);
 }
 
-PyMODINIT_FUNC
-PyInit_simple(void)
-{
-  return module_init("simple", simple);
-}
+PyMODINIT_FUNC PyInit_simple(void) { return module_init("simple", simple); }
